+++ backend/db/src/chapters.rs
@@ -61,7 +61,7 @@ pub async fn get_all_chapter_data() -> Result<Vec<ChapterEntryForEditor>> {
     let chapter_data = sqlx::query_as!(
         DatabaseRow,
         r#"SELECT id, name, desc_sv, desc_en, public,
-            (created_at >= NOW() - $1::interval AND created_at >= DATE '2026-08-17') AS "is_new!"
+            (created_at >= NOW() - $1::interval) AS "is_new!"
             FROM chapters ORDER BY name"#,
         NEW_THRESHOLD
     )
@@ -83,7 +83,7 @@ pub async fn get_public_chapters_from_course(course_id: &i32) -> Result<Vec<Chap
     let chapters = sqlx::query_as!(
         DatabaseRow,
         r#"SELECT ch.id, ch.name, ch.desc_sv, ch.desc_en, ch.public,
-            (created_at >= NOW() - $3::interval AND created_at >= DATE '2026-08-17') AS "is_new!"
+            (created_at >= NOW() - $3::interval) AS "is_new!"
         FROM chapters ch
         JOIN course_chapters cc ON ch.id = cc.chapter_id
         WHERE cc.course_id = $1
@@ -109,7 +109,7 @@ pub async fn get_all_chapters_from_course(course_id: &i32) -> Result<Vec<Chapter
     let chapters = sqlx::query_as!(
         DatabaseRow,
         r#"SELECT ch.id, ch.name, ch.desc_sv, ch.desc_en, ch.public,
-            (created_at >= NOW() - $3::interval AND created_at >= DATE '2026-08-17') AS "is_new!"
+            (created_at >= NOW() - $3::interval) AS "is_new!"
         FROM chapters ch
         JOIN course_chapters cc ON ch.id = cc.chapter_id
         WHERE cc.course_id = $1
@@ -162,7 +162,7 @@ pub async fn get_chapters_from_topic(topic_id: &i32) -> Result<Vec<ChapterEntryF
     let chapters = sqlx::query_as!(
         DatabaseRow,
         r#"SELECT c.id, c.name, c.desc_sv, c.desc_en, c.public,
-            (created_at >= NOW() - $2::interval AND created_at >= DATE '2026-08-17') AS "is_new!"
+            (created_at >= NOW() - $2::interval) AS "is_new!"
         FROM chapters c
         JOIN chapter_topics ct ON c.id = ct.chapter_id
         WHERE ct.topic_id = $1